requirements.txt
absl-py
easydict
cython
numpy
tensorflow-gpu
tqdm
